Added version 5.8.2978

+++ chromium/components/services/storage/shared_storage/shared_storage_database.cc
@@ -15,6 +15,7 @@
 
 #include "base/files/file_util.h"
 #include "base/metrics/histogram_functions.h"
+#include "base/strings/utf_string_conversions.h"
 #include "base/time/default_clock.h"
 #include "base/time/time.h"
 #include "components/services/storage/public/mojom/storage_usage_info.mojom.h"
@@ -287,10 +288,6 @@ SharedStorageDatabase::OperationResult SharedStorageDatabase::Set(
   if (LazyInit(DBCreationPolicy::kCreateIfAbsent) != InitStatus::kSuccess)
     return OperationResult::kInitFailure;
 
-  sql::Transaction transaction(&db_);
-  if (!transaction.Begin())
-    return OperationResult::kSqlError;
-
   GetResult get_result = Get(context_origin, key);
   if (get_result.result != OperationResult::kSuccess &&
       get_result.result != OperationResult::kNotFound &&
@@ -299,30 +296,24 @@ SharedStorageDatabase::OperationResult SharedStorageDatabase::Set(
   }
 
   std::string origin_str(SerializeOrigin(context_origin));
-  if (get_result.result == OperationResult::kSuccess ||
-      get_result.result == OperationResult::kExpired) {
-    if (Delete(context_origin, key) != OperationResult::kSuccess)
+  if (get_result.result == OperationResult::kSuccess &&
+      behavior == SharedStorageDatabase::SetBehavior::kIgnoreIfPresent) {
+    // We re-insert the old key-value pair with an updated `last_used_time`.
+    if (!UpdateValuesMapping(origin_str, key, get_result.data,
+                             /*key_exists=*/true)) {
       return OperationResult::kSqlError;
-
-    if (behavior == SharedStorageDatabase::SetBehavior::kIgnoreIfPresent &&
-        get_result.result != OperationResult::kExpired) {
-      // We re-insert the old key-value pair with an updated `last_used_time`.
-      if (!InsertIntoValuesMapping(origin_str, key, get_result.data))
-        return OperationResult::kSqlError;
-
-      if (!transaction.Commit())
-        return OperationResult::kSqlError;
-      return OperationResult::kIgnored;
     }
-  } else if (!HasCapacity(origin_str)) {
+    return OperationResult::kIgnored;
+  }
+  if (get_result.result == OperationResult::kNotFound &&
+      !HasCapacity(origin_str)) {
     return OperationResult::kNoCapacity;
   }
 
-  if (!InsertIntoValuesMapping(origin_str, key, value))
+  bool key_exists = get_result.result != OperationResult::kNotFound;
+  if (!UpdateValuesMapping(origin_str, key, value, key_exists)) {
     return OperationResult::kSqlError;
-
-  if (!transaction.Commit())
-    return OperationResult::kSqlError;
+  }
 
   return OperationResult::kSet;
 }
@@ -338,10 +329,6 @@ SharedStorageDatabase::OperationResult SharedStorageDatabase::Append(
 
   if (LazyInit(DBCreationPolicy::kCreateIfAbsent) != InitStatus::kSuccess)
     return OperationResult::kInitFailure;
-
-  sql::Transaction transaction(&db_);
-  if (!transaction.Begin())
-    return OperationResult::kSqlError;
 
   GetResult get_result = Get(context_origin, key);
   if (get_result.result != OperationResult::kSuccess &&
@@ -359,14 +346,8 @@ SharedStorageDatabase::OperationResult SharedStorageDatabase::Append(
 
     if (new_value.size() > max_string_length_)
       return OperationResult::kInvalidAppend;
-
-    if (Delete(context_origin, key) != OperationResult::kSuccess)
-      return OperationResult::kSqlError;
   } else if (get_result.result == OperationResult::kExpired) {
     new_value = std::move(tail_value);
-
-    if (Delete(context_origin, key) != OperationResult::kSuccess)
-      return OperationResult::kSqlError;
   } else {
     new_value = std::move(tail_value);
 
@@ -374,11 +355,10 @@ SharedStorageDatabase::OperationResult SharedStorageDatabase::Append(
       return OperationResult::kNoCapacity;
   }
 
-  if (!InsertIntoValuesMapping(origin_str, key, new_value))
+  bool key_exists = get_result.result != OperationResult::kNotFound;
+  if (!UpdateValuesMapping(origin_str, key, new_value, key_exists)) {
     return OperationResult::kSqlError;
-
-  if (!transaction.Commit())
-    return OperationResult::kSqlError;
+  }
 
   return OperationResult::kSet;
 }
@@ -730,9 +710,9 @@ SharedStorageDatabase::PurgeMatchingOrigins(
 
   for (const auto& origin : origins) {
     if (storage_key_matcher &&
-        !storage_key_matcher.Run(
-            blink::StorageKey(url::Origin::Create(GURL(origin))),
-            special_storage_policy_.get())) {
+        !storage_key_matcher.Run(blink::StorageKey::CreateFirstParty(
+                                     url::Origin::Create(GURL(origin))),
+                                 special_storage_policy_.get())) {
       continue;
     }
 
@@ -766,57 +746,43 @@ SharedStorageDatabase::OperationResult SharedStorageDatabase::PurgeStale() {
   if (!transaction.Begin())
     return OperationResult::kSqlError;
 
+  static constexpr char kUpdateLengthsSql[] =
+      "UPDATE per_origin_mapping SET length = length - counts.num_expired "
+      "FROM "
+      "    (SELECT context_origin, COUNT(context_origin) AS num_expired "
+      "    FROM values_mapping WHERE last_used_time<? "
+      "    GROUP BY context_origin) "
+      "AS counts "
+      "WHERE per_origin_mapping.context_origin = counts.context_origin";
+
+  sql::Statement update_statement(
+      db_.GetCachedStatement(SQL_FROM_HERE, kUpdateLengthsSql));
+  base::Time cutoff_time = clock_->Now() - staleness_threshold_;
+  update_statement.BindTime(0, cutoff_time);
+
+  if (!update_statement.Run()) {
+    return OperationResult::kSqlError;
+  }
+
   static constexpr char kDeleteEntriesSql[] =
       "DELETE FROM values_mapping WHERE last_used_time<?";
   sql::Statement entries_statement(
       db_.GetCachedStatement(SQL_FROM_HERE, kDeleteEntriesSql));
-  entries_statement.BindTime(0, clock_->Now() - staleness_threshold_);
+  entries_statement.BindTime(0, cutoff_time);
 
   // Delete expired entries.
   if (!entries_statement.Run())
     return OperationResult::kSqlError;
 
-  static constexpr char kSelectSql[] =
-      "SELECT context_origin,creation_time,length FROM per_origin_mapping "
-      "ORDER BY creation_time";
-  sql::Statement select_statement(
-      db_.GetCachedStatement(SQL_FROM_HERE, kSelectSql));
+  static constexpr char kDeleteOriginsSql[] =
+      "DELETE FROM per_origin_mapping WHERE length<=0";
+  sql::Statement origins_statement(
+      db_.GetCachedStatement(SQL_FROM_HERE, kDeleteOriginsSql));
 
-  // Update `per_origin_mapping` to agree with the correct post-purge `length`s.
-  // Also, for any origins that are now empty and whose `creation_time`s are
-  // prior to the lookback window, remove them from `per_origin_mapping`
-  // entirely.
-  while (select_statement.Step()) {
-    std::string origin = select_statement.ColumnString(0);
-    base::Time creation_time = select_statement.ColumnTime(1);
-    int64_t reported_length = select_statement.ColumnInt64(2);
-
-    // We want the count of all entries for this origin. There are unlikely to
-    // be any expired ones, since we just purged them, but one or more
-    // additional unpurged entries could have become expired in the meantime.
-    int64_t actual_length =
-        NumEntriesManualCount(origin, /*include_expired=*/true);
-
-    if (actual_length == -1)
-      return OperationResult::kSqlError;
-
-    if (actual_length == 0 &&
-        creation_time < clock_->Now() - staleness_threshold_) {
-      if (!DeleteFromPerOriginMapping(origin))
-        return OperationResult::kSqlError;
-      continue;
-    }
-
-    if (actual_length == reported_length) {
-      continue;
-    }
-
-    if (!UpdateLength(origin, actual_length - reported_length))
-      return OperationResult::kSqlError;
-  }
-
-  if (!select_statement.Succeeded())
+  // Delete empty origins.
+  if (!origins_statement.Run()) {
     return OperationResult::kSqlError;
+  }
 
   static constexpr char kDeleteWithdrawalsSql[] =
       "DELETE FROM budget_mapping WHERE time_stamp<?";
@@ -834,27 +800,24 @@ SharedStorageDatabase::OperationResult SharedStorageDatabase::PurgeStale() {
   return OperationResult::kSuccess;
 }
 
-std::vector<mojom::StorageUsageInfoPtr> SharedStorageDatabase::FetchOrigins(
-    bool exclude_empty_origins) {
+std::vector<mojom::StorageUsageInfoPtr> SharedStorageDatabase::FetchOrigins() {
   DCHECK_CALLED_ON_VALID_SEQUENCE(sequence_checker_);
 
   if (LazyInit(DBCreationPolicy::kIgnoreIfAbsent) != InitStatus::kSuccess)
     return {};
 
-  const char* kSelectSql = (exclude_empty_origins)
-                               ? "SELECT context_origin,creation_time,length "
-                                 "FROM per_origin_mapping "
-                                 "WHERE length>0 ORDER BY context_origin"
-                               : "SELECT context_origin,creation_time,length "
-                                 "FROM per_origin_mapping "
-                                 "ORDER BY context_origin";
+  static constexpr char kSelectSql[] =
+      "SELECT context_origin,creation_time,length "
+      "FROM per_origin_mapping "
+      "ORDER BY context_origin";
 
-  sql::Statement statement(db_.GetUniqueStatement(kSelectSql));
+  sql::Statement statement(db_.GetCachedStatement(SQL_FROM_HERE, kSelectSql));
   std::vector<mojom::StorageUsageInfoPtr> fetched_origin_infos;
 
   while (statement.Step()) {
     fetched_origin_infos.emplace_back(mojom::StorageUsageInfo::New(
-        blink::StorageKey(url::Origin::Create(GURL(statement.ColumnString(0)))),
+        blink::StorageKey::CreateFirstParty(
+            url::Origin::Create(GURL(statement.ColumnString(0)))),
         statement.ColumnInt64(2) * kSharedStorageEntryTotalBytesMultiplier *
             max_string_length_,
         statement.ColumnTime(1)));
@@ -1002,6 +965,32 @@ SharedStorageDatabase::GetEntriesForDevTools(url::Origin context_origin) {
   return entries;
 }
 
+SharedStorageDatabase::OperationResult
+SharedStorageDatabase::ResetBudgetForDevTools(url::Origin context_origin) {
+  DCHECK_CALLED_ON_VALID_SEQUENCE(sequence_checker_);
+
+  if (LazyInit(DBCreationPolicy::kIgnoreIfAbsent) != InitStatus::kSuccess) {
+    // We do not return an error if the database doesn't exist, but only if it
+    // pre-exists on disk and yet fails to initialize.
+    if (db_status_ == InitStatus::kUnattempted) {
+      return OperationResult::kSuccess;
+    } else {
+      return OperationResult::kInitFailure;
+    }
+  }
+
+  static constexpr char kDeleteSql[] =
+      "DELETE FROM budget_mapping WHERE context_origin=?";
+
+  sql::Statement statement(db_.GetCachedStatement(SQL_FROM_HERE, kDeleteSql));
+  statement.BindString(0, SerializeOrigin(context_origin));
+
+  if (!statement.Run()) {
+    return OperationResult::kSqlError;
+  }
+  return OperationResult::kSuccess;
+}
+
 bool SharedStorageDatabase::IsOpenForTesting() const {
   DCHECK_CALLED_ON_VALID_SEQUENCE(sequence_checker_);
   return db_.is_open();
@@ -1036,8 +1025,8 @@ bool SharedStorageDatabase::OverrideCreationTimeForTesting(
   if (result == OperationResult::kNotFound)
     return true;
 
-  return DeleteThenMaybeInsertIntoPerOriginMapping(
-      origin_str, new_creation_time, length, /*force_insertion=*/true);
+  return UpdatePerOriginMapping(origin_str, new_creation_time, length,
+                                /*origin_exists=*/true);
 }
 
 bool SharedStorageDatabase::OverrideLastUsedTimeForTesting(
@@ -1059,11 +1048,9 @@ bool SharedStorageDatabase::OverrideLastUsedTimeForTesting(
   if (result.result == OperationResult::kNotFound)
     return true;
 
-  if (Delete(context_origin, key) != OperationResult::kSuccess)
-    return false;
-
-  if (!InsertIntoValuesMappingWithTime(SerializeOrigin(context_origin), key,
-                                       result.data, new_last_used_time)) {
+  if (!UpdateValuesMappingWithTime(SerializeOrigin(context_origin), key,
+                                   result.data, new_last_used_time,
+                                   /*key_exists=*/true)) {
     return false;
   }
   return true;
@@ -1331,8 +1318,12 @@ bool SharedStorageDatabase::Vacuum() {
   return db_.Execute("VACUUM");
 }
 
-bool SharedStorageDatabase::Purge(const std::string& context_origin,
-                                  bool delete_origin_if_empty) {
+bool SharedStorageDatabase::Purge(const std::string& context_origin) {
+  sql::Transaction transaction(&db_);
+  if (!transaction.Begin()) {
+    return false;
+  }
+
   static constexpr char kDeleteSql[] =
       "DELETE FROM values_mapping "
       "WHERE context_origin=?";
@@ -1340,29 +1331,10 @@ bool SharedStorageDatabase::Purge(const std::string& context_origin,
   sql::Statement statement(db_.GetCachedStatement(SQL_FROM_HERE, kDeleteSql));
   statement.BindString(0, context_origin);
 
-  sql::Transaction transaction(&db_);
-  if (!transaction.Begin())
-    return false;
-
   if (!statement.Run())
     return false;
 
-  int64_t length = 0L;
-  base::Time creation_time;
-  OperationResult result =
-      GetOriginInfo(context_origin, &length, &creation_time);
-
-  if (result != OperationResult::kSuccess &&
-      result != OperationResult::kNotFound) {
-    return false;
-  }
-
-  // Don't delete or insert for non-existent origin.
-  if (result == OperationResult::kNotFound)
-    return true;
-
-  if (!DeleteThenMaybeInsertIntoPerOriginMapping(context_origin, creation_time,
-                                                 0L, !delete_origin_if_empty)) {
+  if (!DeleteFromPerOriginMapping(context_origin)) {
     return false;
   }
 
@@ -1390,19 +1362,14 @@ int64_t SharedStorageDatabase::NumEntriesTotal(
 }
 
 int64_t SharedStorageDatabase::NumEntriesManualCount(
-    const std::string& context_origin,
-    bool include_expired) {
-  const char* kCountSql =
-      (include_expired)
-          ? "SELECT COUNT(*) FROM values_mapping WHERE context_origin=?"
-          : "SELECT COUNT(*) FROM values_mapping "
-            "WHERE context_origin=? AND last_used_time>=?";
+    const std::string& context_origin) {
+  static constexpr char kCountSql[] =
+      "SELECT COUNT(*) FROM values_mapping "
+      "WHERE context_origin=? AND last_used_time>=?";
 
-  sql::Statement statement(db_.GetUniqueStatement(kCountSql));
+  sql::Statement statement(db_.GetCachedStatement(SQL_FROM_HERE, kCountSql));
   statement.BindString(0, context_origin);
-
-  if (!include_expired)
-    statement.BindTime(1, clock_->Now() - staleness_threshold_);
+  statement.BindTime(1, clock_->Now() - staleness_threshold_);
 
   int64_t length = 0;
   if (statement.Step())
@@ -1459,8 +1426,7 @@ SharedStorageDatabase::OperationResult SharedStorageDatabase::GetOriginInfo(
 }
 
 bool SharedStorageDatabase::UpdateLength(const std::string& context_origin,
-                                         int64_t delta,
-                                         bool delete_origin_if_empty) {
+                                         int64_t delta) {
   int64_t length = 0L;
   base::Time creation_time;
   OperationResult result =
@@ -1471,6 +1437,7 @@ bool SharedStorageDatabase::UpdateLength(const std::string& context_origin,
     return false;
   }
 
+  bool origin_exists = true;
   if (result == OperationResult::kNotFound) {
     // Don't delete or insert for non-existent origin when we would have
     // decremented the length.
@@ -1479,25 +1446,41 @@ bool SharedStorageDatabase::UpdateLength(const std::string& context_origin,
 
     // We are creating `context_origin` now.
     creation_time = clock_->Now();
+    origin_exists = false;
   }
 
   int64_t new_length = (length + delta > 0L) ? length + delta : 0L;
 
-  return DeleteThenMaybeInsertIntoPerOriginMapping(
-      context_origin, creation_time, new_length, !delete_origin_if_empty);
+  return UpdatePerOriginMapping(context_origin, creation_time, new_length,
+                                origin_exists);
 }
 
-bool SharedStorageDatabase::InsertIntoValuesMappingWithTime(
+bool SharedStorageDatabase::UpdateValuesMappingWithTime(
     const std::string& context_origin,
     const std::u16string& key,
     const std::u16string& value,
-    base::Time last_used_time) {
+    base::Time last_used_time,
+    bool key_exists) {
+  if (key_exists) {
+    static constexpr char kUpdateSql[] =
+        "UPDATE values_mapping SET value=?, last_used_time=? "
+        "WHERE context_origin=? AND key=?";
+
+    sql::Statement statement(db_.GetCachedStatement(SQL_FROM_HERE, kUpdateSql));
+    statement.BindString16(0, value);
+    statement.BindTime(1, last_used_time);
+    statement.BindString(2, context_origin);
+    statement.BindString16(3, key);
+
+    return statement.Run();
+  }
+
   sql::Transaction transaction(&db_);
   if (!transaction.Begin())
     return false;
 
   static constexpr char kInsertSql[] =
-      "INSERT INTO values_mapping(context_origin,key,value,last_used_time)"
+      "INSERT INTO values_mapping(context_origin,key,value,last_used_time) "
       "VALUES(?,?,?,?)";
 
   sql::Statement statement(db_.GetCachedStatement(SQL_FROM_HERE, kInsertSql));
@@ -1515,12 +1498,13 @@ bool SharedStorageDatabase::InsertIntoValuesMappingWithTime(
   return transaction.Commit();
 }
 
-bool SharedStorageDatabase::InsertIntoValuesMapping(
+bool SharedStorageDatabase::UpdateValuesMapping(
     const std::string& context_origin,
     const std::u16string& key,
-    const std::u16string& value) {
-  return InsertIntoValuesMappingWithTime(context_origin, key, value,
-                                         clock_->Now());
+    const std::u16string& value,
+    bool key_exists) {
+  return UpdateValuesMappingWithTime(context_origin, key, value, clock_->Now(),
+                                     key_exists);
 }
 
 bool SharedStorageDatabase::DeleteFromPerOriginMapping(
@@ -1540,7 +1524,7 @@ bool SharedStorageDatabase::InsertIntoPerOriginMapping(
     base::Time creation_time,
     uint64_t length) {
   static constexpr char kInsertSql[] =
-      "INSERT INTO per_origin_mapping(context_origin,creation_time,length)"
+      "INSERT INTO per_origin_mapping(context_origin,creation_time,length) "
       "VALUES(?,?,?)";
 
   sql::Statement statement(db_.GetCachedStatement(SQL_FROM_HERE, kInsertSql));
@@ -1551,26 +1535,34 @@ bool SharedStorageDatabase::InsertIntoPerOriginMapping(
   return statement.Run();
 }
 
-bool SharedStorageDatabase::DeleteThenMaybeInsertIntoPerOriginMapping(
+bool SharedStorageDatabase::UpdatePerOriginMapping(
     const std::string& context_origin,
     base::Time creation_time,
     uint64_t length,
-    bool force_insertion) {
+    bool origin_exists) {
   DCHECK(length >= 0L);
 
-  sql::Transaction transaction(&db_);
-  if (!transaction.Begin())
-    return false;
+  if (length && origin_exists) {
+    static constexpr char kUpdateSql[] =
+        "UPDATE per_origin_mapping SET creation_time=?, length=? "
+        "WHERE context_origin=?";
+    sql::Statement statement(db_.GetCachedStatement(SQL_FROM_HERE, kUpdateSql));
+    statement.BindTime(0, creation_time);
+    statement.BindInt64(1, static_cast<int64_t>(length));
+    statement.BindString(2, context_origin);
 
-  if (!DeleteFromPerOriginMapping(context_origin))
-    return false;
-
-  if ((length || force_insertion) &&
-      !InsertIntoPerOriginMapping(context_origin, creation_time, length)) {
-    return false;
+    return statement.Run();
+  }
+  if (length) {
+    return InsertIntoPerOriginMapping(context_origin, creation_time, length);
+  }
+  if (origin_exists) {
+    return DeleteFromPerOriginMapping(context_origin);
   }
 
-  return transaction.Commit();
+  //  Origin does not exist and we are trying to set the length to 0, so this is
+  //  a no-op.
+  return true;
 }
 
 bool SharedStorageDatabase::HasCapacity(const std::string& context_origin) {
